Add files via upload
During the data visualization process of the given data set we found that there were 4000 rows and 30 columns. We also looked at the meta data and made sure to understand the significance of the variables, also tried to find all the relationships between all the data attributes them by plotting to be able to improve the sales. This study aims to determine how this company can increase its profits. We did data analysis on sales raw data, started filtering to focus on the important data, then mapped the data. The results of our analysis will also be shown in the report.